# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type SourceStatus = typeof SourceStatus[keyof typeof SourceStatus];


export const SourceStatus = {
  live: 'live',
  cached: 'cached',
} as const;

export interface Source {
  name: string;
  url: string;
  fetchedAt: string;
  status: SourceStatus;
}

export interface Team {
  /** @nullable */
  id: number | null;
  name: string;
  /** @nullable */
  crestUrl: string | null;
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  upcoming: 'upcoming',
  live: 'live',
  finished: 'finished',
  postponed: 'postponed',
} as const;

export interface Match {
  id: string;
  matchId: number;
  slug: string;
  competition: string;
  /** @nullable */
  competitionId?: number | null;
  /** @nullable */
  round?: string | null;
  /** @nullable */
  kickoff?: string | null;
  /** @nullable */
  kickoffText?: string | null;
  /** @nullable */
  venue?: string | null;
  statusText?: string;
  status: MatchStatus;
  homeTeam: Team;
  awayTeam: Team;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  url: string;
}

export interface MatchesResponse {
  matches: Match[];
  source: Source;
}

export interface LineupPlayer {
  id: number;
  name: string;
  /** @nullable */
  number?: number | null;
  position: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  minutesPlayed?: number | null;
  isCaptain: boolean;
  isSpare: boolean;
}

export interface MatchEvent {
  id: number;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  addedTime?: number | null;
  type: string;
  /** @nullable */
  half?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  player?: string | null;
  /** @nullable */
  playerPhotoUrl?: string | null;
  /** @nullable */
  relatedPlayer?: string | null;
}

export type MatchDetailLineups = {
  home: LineupPlayer[];
  away: LineupPlayer[];
  homeBench: LineupPlayer[];
  awayBench: LineupPlayer[];
};

export type MatchDetail = Match & ({
  /** @nullable */
  referee?: string | null;
  /** @nullable */
  stadium?: string | null;
  /** @nullable */
  homeCoach?: string | null;
  /** @nullable */
  awayCoach?: string | null;
  /** @nullable */
  homeFormation?: string | null;
  /** @nullable */
  awayFormation?: string | null;
  tvChannels: string[];
  events: MatchEvent[];
  timeline: MatchEvent[];
  lineups: MatchDetailLineups;
});

export interface MatchDetailResponse {
  match: MatchDetail;
  source: Source;
}

export interface Player {
  id: number;
  name: string;
  /** @nullable */
  number?: number | null;
  position: string;
  nationality: string;
  /** @nullable */
  photoUrl?: string | null;
  url: string;
  /** @nullable */
  goals?: number | null;
  /** @nullable */
  appearances?: number | null;
  /** @nullable */
  scoringRate?: number | null;
}

export type SquadResponseCoach = { [key: string]: unknown };

export type SquadResponseScorersItem = { [key: string]: unknown };

export interface SquadResponse {
  players: Player[];
  coach: SquadResponseCoach;
  scorers: SquadResponseScorersItem[];
  source: Source;
}

export type TeamResponseCoach = { [key: string]: unknown };

export type TeamResponseScorersItem = { [key: string]: unknown };

export interface TeamResponse {
  team: Team;
  players: Player[];
  coach: TeamResponseCoach;
  scorers: TeamResponseScorersItem[];
  source: Source;
}

export interface Standing {
  rank: number;
  team: Team;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  points: number;
  isMasry: boolean;
}

export interface StandingsResponse {
  standings: Standing[];
  source: Source;
}

export interface NewsItem {
  id: string;
  title: string;
  url: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  publishedText?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  sourceName: string;
}

export interface NewsResponse {
  news: NewsItem[];
  source: Source;
}

export interface HeadToHeadMeeting { [key: string]: unknown }

export type HeadToHeadResponseOpponent = {
  id: number;
  name: string;
};

export type HeadToHeadResponseSummary = { [key: string]: unknown };

/**
 * @nullable
 */
export type HeadToHeadResponseTopScorer = { [key: string]: unknown } | null;

export interface HeadToHeadResponse {
  opponent: HeadToHeadResponseOpponent;
  source: Source;
  summary: HeadToHeadResponseSummary;
  /** @nullable */
  topScorer?: HeadToHeadResponseTopScorer;
  meetings: HeadToHeadMeeting[];
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Source provider unavailable
 */
export type UpstreamUnavailableResponse = ErrorResponse;

export type GetHeadToHeadParams = {
opponent: string;
};

export type GetPlayerDetail200 = {
  player: Player;
  source: Source;
};

export type RedirectFootballImageParams = {
url: string;
};