# lib/api-zod/src/generated/types/headToHeadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { HeadToHeadMeeting } from './headToHeadMeeting';
import type { HeadToHeadResponseOpponent } from './headToHeadResponseOpponent';
import type { HeadToHeadResponseSummary } from './headToHeadResponseSummary';
import type { HeadToHeadResponseTopScorer } from './headToHeadResponseTopScorer';
import type { Source } from './source';

export interface HeadToHeadResponse {
  opponent: HeadToHeadResponseOpponent;
  source: Source;
  summary: HeadToHeadResponseSummary;
  /** @nullable */
  topScorer?: HeadToHeadResponseTopScorer;
  meetings: HeadToHeadMeeting[];
}