# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { Match } from './match';
import type { MatchDetailLineups } from './matchDetailLineups';
import type { MatchEvent } from './matchEvent';

export type MatchDetail = Match & ({
  /** @nullable */
  referee?: string | null;
  /** @nullable */
  stadium?: string | null;
  /** @nullable */
  homeCoach?: string | null;
  /** @nullable */
  awayCoach?: string | null;
  /** @nullable */
  homeFormation?: string | null;
  /** @nullable */
  awayFormation?: string | null;
  tvChannels: string[];
  events: MatchEvent[];
  timeline: MatchEvent[];
  lineups: MatchDetailLineups;
});