# lib/api-zod/src/generated/types/matchEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */

export interface MatchEvent {
  id: number;
  /** @nullable */
  minute?: number | null;
  /** @nullable */
  addedTime?: number | null;
  type: string;
  /** @nullable */
  half?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  player?: string | null;
  /** @nullable */
  playerPhotoUrl?: string | null;
  /** @nullable */
  relatedPlayer?: string | null;
}