# lib/api-zod/src/generated/types/matchesResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { Match } from './match';
import type { Source } from './source';

export interface MatchesResponse {
  matches: Match[];
  source: Source;
}