# lib/api-zod/src/generated/types/newsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { NewsItem } from './newsItem';
import type { Source } from './source';

export interface NewsResponse {
  news: NewsItem[];
  source: Source;
}