# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */

export interface Player {
  id: number;
  name: string;
  /** @nullable */
  number?: number | null;
  position: string;
  nationality: string;
  /** @nullable */
  photoUrl?: string | null;
  url: string;
  /** @nullable */
  goals?: number | null;
  /** @nullable */
  appearances?: number | null;
  /** @nullable */
  scoringRate?: number | null;
}