# lib/api-zod/src/generated/types/standingsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { Source } from './source';
import type { Standing } from './standing';

export interface StandingsResponse {
  standings: Standing[];
  source: Source;
}