# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */

export interface Team {
  /** @nullable */
  id: number | null;
  name: string;
  /** @nullable */
  crestUrl: string | null;
}