# lib/api-zod/src/generated/types/teamResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public football data API for Egyptian Football Icons
 * OpenAPI spec version: 1.0.0
 */
import type { Player } from './player';
import type { Source } from './source';
import type { Team } from './team';
import type { TeamResponseCoach } from './teamResponseCoach';
import type { TeamResponseScorersItem } from './teamResponseScorersItem';

export interface TeamResponse {
  team: Team;
  players: Player[];
  coach: TeamResponseCoach;
  scorers: TeamResponseScorersItem[];
  source: Source;
}